pyproject.toml
[project]
name = "ai-research-assistant-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen>=0.11.2",
    "chromadb>=1.0.20",
    "flask>=3.1.3",
    "groq>=1.0.0",
    "langchain-text-splitters>=0.3.11",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
]
